load LCStep with blank input field
